requirements.txt
Django==2.2.2
djangorestframework==3.9.4
request==2019.4.13
requests==2.22.0
urllib3==1.25.3
httpie==1.0.2